refactor : 차단 유저 조회 쿼리문 도메인 변경
