+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/block/repository/BlockRepository.java
@@ -1,7 +1,11 @@
 package org.j1p5.domain.block.repository;
 
 import org.j1p5.domain.block.entity.BlockEntity;
-import org.j1p5.domain.block.repository.querydsl.BlockRepositoryCustom;
+import org.j1p5.domain.user.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlockRepository extends JpaRepository<BlockEntity, Long>, BlockRepositoryCustom {}
+import java.util.List;
+
+public interface BlockRepository extends JpaRepository<BlockEntity, Long> {
+    List<BlockEntity> findByUser(UserEntity user);
+}
