--- meerket/meerket-domain/src/main/java/org/j1p5/domain/block/repository/querydsl/BlockRepositoryCustom.java
@@ -1,12 +1,0 @@
-package org.j1p5.domain.block.repository.querydsl;
-
-import org.j1p5.domain.block.BlockUserInfo;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface BlockRepositoryCustom {
-    Page<BlockUserInfo> getBlockUsers(Long userId, Pageable pageable);
-
-
-
-}
