+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/user/repository/custom/UserRepositoryCustomImpl.java
@@ -1,35 +1,31 @@
-package org.j1p5.domain.block.repository.querydsl;
+package org.j1p5.domain.user.repository.custom;
 
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import org.j1p5.domain.activityArea.dto.ActivityAreaAddress;
 import org.j1p5.domain.activityArea.entity.QActivityArea;
 import org.j1p5.domain.block.BlockUserInfo;
-import org.j1p5.domain.block.entity.QBlockEntity;
 import org.j1p5.domain.user.entity.QEmdArea;
 import org.j1p5.domain.user.entity.QUserEntity;
-import org.locationtech.jts.geom.Point;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
-public class BlockRepositoryCustomImpl implements BlockRepositoryCustom {
+public class UserRepositoryCustomImpl implements UserRepositoryCustom {
 
     private final JPAQueryFactory jpaQueryFactory;
 
-    public BlockRepositoryCustomImpl(JPAQueryFactory jpaQueryFactory) {
+    public UserRepositoryCustomImpl(JPAQueryFactory jpaQueryFactory) {
         this.jpaQueryFactory = jpaQueryFactory;
     }
 
-    QBlockEntity qBlockEntity = QBlockEntity.blockEntity;
     QUserEntity qUserEntity = QUserEntity.userEntity;
     QEmdArea qEmdArea = QEmdArea.emdArea;
     QActivityArea qActivityArea = QActivityArea.activityArea;
 
     @Override
-    public Page<BlockUserInfo> getBlockUsers(Long userId, Pageable pageable) {
+    public Page<BlockUserInfo> findBlockUserByIds(List<Long> userIds, Pageable pageable) {
         List<BlockUserInfo> blockUserInfos =
                 jpaQueryFactory
                         .selectDistinct(
@@ -41,12 +37,11 @@ public class BlockRepositoryCustomImpl implements BlockRepositoryCustom {
                                         qEmdArea.emdName.as("emdName")
                                 )
                         )
-                        .from(qBlockEntity)
+                        .from(qUserEntity)
                         //TODO : 역정규화를 사용한 리팩토링(성능개선) 필요
-                        .join(qUserEntity).on(qBlockEntity.user.eq(qUserEntity))
                         .join(qActivityArea).on(qActivityArea.user.eq(qUserEntity))
                         .join(qEmdArea).on(qActivityArea.emdArea.eq(qEmdArea))
-                        .where(qBlockEntity.user.id.eq(userId))
+                        .where(qUserEntity.id.in(userIds))
                         .offset(pageable.getOffset())
                         .limit(pageable.getPageSize())
                         .fetch();
@@ -56,12 +51,11 @@ public class BlockRepositoryCustomImpl implements BlockRepositoryCustom {
         }
 
         Long totalCount = jpaQueryFactory
-                .selectDistinct(qBlockEntity.count())
-                .from(qBlockEntity)
-                .join(qUserEntity).on(qBlockEntity.user.eq(qUserEntity))
+                .selectDistinct(qUserEntity.count())
+                .from(qUserEntity)
                 .join(qActivityArea).on(qActivityArea.user.eq(qUserEntity))
                 .join(qEmdArea).on(qActivityArea.emdArea.eq(qEmdArea))
-                .where(qBlockEntity.user.id.eq(userId))
+                .where(qUserEntity.id.in(userIds))
                 .fetchOne();
 
         if (totalCount == null) {
